+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 0.2.8
+
+- add Linter, cleanup error messages
+- add status checks
+- update dependencies
+- now also shows unchanged but touched traits
+
 ## 0.2.7
 
 - upgrade docs